upd: move CVerifyDb from main
