+++ divi/src/init.cpp
@@ -71,6 +71,9 @@
 #ifndef BITCOIN_VALIDATIONSTATE_H
 #include <ValidationState.h>
 #endif
+#ifndef BITCOIN_VERIFYDB_H
+#include <verifyDb.h>
+#endif
 
 #ifdef ENABLE_WALLET
 CWallet* pwalletMain = NULL;
