+++ divi/src/rpcblockchain.cpp
@@ -20,6 +20,9 @@
 #ifndef BITCOIN_VALIDATIONSTATE_H
 #include <ValidationState.h>
 #endif
+#ifndef BITCOIN_VERIFYDB_H
+#include <verifyDb.h>
+#endif
 
 using namespace json_spirit;
 using namespace std;
