+++ divi/src/verifyDb.h
@@ -1,0 +1,32 @@
+// Copyright (c) 2009-2010 Satoshi Nakamoto
+// Copyright (c) 2009-2014 The Bitcoin developers
+// Copyright (c) 2014-2015 The Dash developers
+// Copyright (c) 2015-2017 The PIVX Developers
+// Copyright (c) 2017-2020 The DIVI Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+#ifndef BITCOIN_VERIFYDB_H
+#define BITCOIN_VERIFYDB_H
+
+#ifndef BITCOIN_COINS_H
+#include <coins.h>
+#endif
+
+#ifndef LOGGING_H
+#include <Logging.h>
+#endif
+
+#ifndef BITCOIN_CHAIN_H
+#include <chain.h>
+#endif
+
+/** RAII wrapper for VerifyDB: Verify consistency of the block and coin databases */
+class CVerifyDB
+{
+public:
+    CVerifyDB();
+    ~CVerifyDB();
+    bool VerifyDB(CCoinsView* coinsview, int nCheckLevel, int nCheckDepth);
+};
+
+#endif
